K105: Reorganized the complex and hypercomplex types more comprehensibly, cleaned up, and added documentation.
