--- kosmos-core/src/main/kotlin/org/vorpal/kosmos/algebra/structures/instances/HyperComplex.kt
@@ -1,13 +1,0 @@
-package org.vorpal.kosmos.algebra.structures.instances
-
-object HyperComplex {
-    /**
-     * Handedness of a map.
-     * - [RIGHT] indicates that we do not flip along the axis.
-     * - [LEFT] indicates that we flip along the axis.
-     */
-    enum class Handedness {
-        RIGHT,
-        LEFT
-    }
-}
