+++ kosmos-core/src/main/kotlin/org/vorpal/kosmos/hypercomplex/complex/GaussianInt.kt
@@ -1,9 +1,11 @@
-package org.vorpal.kosmos.core.gaussian
+package org.vorpal.kosmos.hypercomplex.complex
 
 import java.math.BigInteger
 
 /**
  * The type for a Gaussian integer, i.e. a [BigInteger] real and complex component.
+ *
+ * Note that this could also be generated using `CayleyDickson` on a [ZCommutativeRing].
  */
 data class GaussianInt(val re: BigInteger, val im: BigInteger) {
     operator fun plus(other: GaussianInt): GaussianInt = GaussianInt(re + other.re, im + other.im)
