+++ kosmos-core/src/main/kotlin/org/vorpal/kosmos/hypercomplex/complex/GaussianIntAlgebras.kt
@@ -1,4 +1,4 @@
-package org.vorpal.kosmos.core.gaussian
+package org.vorpal.kosmos.hypercomplex.complex
 
 import org.vorpal.kosmos.algebra.morphisms.RingMonomorphism
 import org.vorpal.kosmos.algebra.structures.AbelianGroup
@@ -6,10 +6,7 @@ import org.vorpal.kosmos.algebra.structures.CommutativeMonoid
 import org.vorpal.kosmos.algebra.structures.CommutativeRing
 import org.vorpal.kosmos.algebra.structures.HasNormSq
 import org.vorpal.kosmos.algebra.structures.InvolutiveRing
-import org.vorpal.kosmos.algebra.structures.instances.Complex
-import org.vorpal.kosmos.algebra.structures.instances.ComplexAlgebras
 import org.vorpal.kosmos.algebra.structures.instances.IntegerAlgebras
-import org.vorpal.kosmos.algebra.structures.instances.complex
 import org.vorpal.kosmos.bridge.ZModule
 import org.vorpal.kosmos.core.Eq
 import org.vorpal.kosmos.core.Symbols
@@ -20,6 +17,21 @@ import org.vorpal.kosmos.core.ops.LeftAction
 import org.vorpal.kosmos.core.ops.UnaryOp
 import java.math.BigInteger
 
+/**
+ * [GaussianIntAlgebras] contains the algebraic structures over the [GaussianInt] type, as well as the
+ * homomorphisms and [Eq] instances.
+ *
+ * These include:
+ * - [GaussianIntCommutativeRing]: the Gaussian integers.
+ * - [ZModuleGaussianInt]: the two-dimensional vector space of Gaussian integers over the integers.
+ *
+ * We have the following homomorphisms:
+ * - [ZToGaussianIntMonomorphism]: from the integers to the Gaussian integers.
+ * - [GaussianIntToComplexMonomorphism]: from the Gaussian integers to the complex numbers.
+ *
+ * We also have the following [Eq]s:
+ * - [eqGaussianInt]: equality on Gaussian integers.
+ */
 object GaussianIntAlgebras {
     /**
      * Note that the Gaussian integers actually form an integral domain and
