+++ kosmos-core/src/main/kotlin/org/vorpal/kosmos/hypercomplex/embeddings/HyperComplex.kt
@@ -1,0 +1,11 @@
+package org.vorpal.kosmos.hypercomplex.embeddings
+
+/**
+ * Handedness of a map.
+ * - [RIGHT] indicates that we do not flip along the axis.
+ * - [LEFT] indicates that we flip along the axis.
+ */
+enum class Handedness {
+    RIGHT,
+    LEFT
+}
