+++ kosmos-core/src/main/kotlin/org/vorpal/kosmos/hypercomplex/quaternion/QuaternionAlgebras.kt
@@ -1,4 +1,4 @@
-package org.vorpal.kosmos.algebra.structures.instances
+package org.vorpal.kosmos.hypercomplex.quaternion
 
 import org.vorpal.kosmos.algebra.morphisms.RingMonomorphism
 import org.vorpal.kosmos.algebra.structures.CD
@@ -10,37 +10,40 @@ import org.vorpal.kosmos.algebra.structures.InvolutiveRing
 import org.vorpal.kosmos.algebra.structures.Monoid
 import org.vorpal.kosmos.algebra.structures.RealNormedDivisionAlgebra
 import org.vorpal.kosmos.algebra.structures.StarAlgebra
-import org.vorpal.kosmos.algebra.structures.instances.embeddings.AxisSignEmbeddings
+import org.vorpal.kosmos.algebra.structures.instances.RealAlgebras
+import org.vorpal.kosmos.hypercomplex.embeddings.AxisSignEmbeddings
 import org.vorpal.kosmos.core.Eq
 import org.vorpal.kosmos.core.Symbols
 import org.vorpal.kosmos.core.math.Real
 import org.vorpal.kosmos.core.ops.Endo
 import org.vorpal.kosmos.core.ops.LeftAction
 import org.vorpal.kosmos.core.ops.UnaryOp
+import org.vorpal.kosmos.hypercomplex.complex.Complex
+import org.vorpal.kosmos.hypercomplex.complex.ComplexAlgebras
+import org.vorpal.kosmos.hypercomplex.complex.im
+import org.vorpal.kosmos.hypercomplex.complex.re
 import java.math.BigInteger
 import kotlin.isFinite
 
-typealias Quaternion = CD<Complex>
-val Quaternion.w: Real get() = a.re
-val Quaternion.x: Real get() = a.im
-val Quaternion.y: Real get() = b.re
-val Quaternion.z: Real get() = b.im
-
 /**
- * Convenience constructor for a quaternion:
+ * [QuaternionAlgebras] contains the algebraic structures over the [Quaternion] type, as well as the homomorphisms
+ * and [Eq] instances.
  *
- *    complex(w + x i_c), complex(y + z i_c)
+ * These include:
+ * - [QuaternionDivisionRing]: the quaternion division ring.
+ * - [QuaternionVectorSpace]: the two-dimensional vector space of quaternions over the real numbers.
+ * - [quaternionLeftComplexVectorSpace]: the two-dimensional vector space of quaternions over the complex numbers.
+ * - [QuaternionLeftComplexVectorSpacesAll]: a map of all complex embeddings to the vector space that they generate.
+ * - [QuaternionLeftComplexVectorSpaceCanonical]: the canonical vector space of quaternions over the complex numbers.
+ * - [QuaternionStarAlgebra]: the quaternion star algebra.
+ *
+ * We have the following homomorphisms:
+ * - [complexEmbeddingToQuaternion]: the unital embeddings from the complex numbers to the quaternions.
+ *
+ * We also have the following [Eq]s:
+ * - [eqQuaternionStrict]: strict equality on quaternions.
+ * - [eqQuaternion]: approximate equality on quaternions.
  */
-fun quaternion(w: Real,
-               x: Real,
-               y: Real,
-               z: Real
-): Quaternion {
-    val a = complex(w, x)
-    val b = complex(y, z)
-    return Quaternion(a, b)
-}
-
 object QuaternionAlgebras {
 
     object QuaternionDivisionRing:
@@ -51,12 +54,13 @@ object QuaternionAlgebras {
         private val base: NonAssociativeInvolutiveRing<Quaternion> =
             CayleyDickson.usual(ComplexAlgebras.ComplexStarAlgebra)
 
+        override val zero = base.add.identity
+        override val one = base.mul.identity
+
         override val add = base.add
 
-        // base.mul is a NonAssociativeMonoid: for quaternions, it is actually a monoid and
-        // Ring expects an associative monoid, so we wrap.
         override val mul: Monoid<Quaternion> = Monoid.of(
-            identity = base.mul.identity,
+            identity = one,
             op = base.mul.op
         )
 
@@ -71,7 +75,8 @@ object QuaternionAlgebras {
 
             Quaternion(
                 ComplexAlgebras.ComplexRealVectorSpace.leftAction(scale, qc.a),
-                ComplexAlgebras.ComplexRealVectorSpace.leftAction(scale, qc.b))
+                ComplexAlgebras.ComplexRealVectorSpace.leftAction(scale, qc.b)
+            )
         }
 
         override fun fromBigInt(n: BigInteger) =
@@ -82,10 +87,6 @@ object QuaternionAlgebras {
 
         override val normSq: UnaryOp<Quaternion, Real> =
             UnaryOp(Symbols.NORM_SQ_SYMBOL){ q -> mul(q, conj(q)).w }
-
-        // Disambiguate zero and one.
-        override val zero = add.identity
-        override val one = mul.identity
     }
 
     // Scalars: Real, act componentwise on (a, b)
@@ -101,10 +102,38 @@ object QuaternionAlgebras {
         }
     )
 
-    // --------------------------------------------------------------------------
-    // Complex -> Quaternion embeddings (six unital choices) via ComplexEmbedding
-    // --------------------------------------------------------------------------
     private val canonicalEmbedding = AxisSignEmbeddings.AxisSignEmbedding.canonical
+
+    /**
+     * ℍ as a left ℂ-vector space using the chosen embedding.
+     */
+    fun quaternionLeftComplexVectorSpace(
+        emb: AxisSignEmbeddings.AxisSignEmbedding = canonicalEmbedding
+    ): FiniteVectorSpace<Complex, Quaternion> {
+        val embed = complexEmbeddingToQuaternion(emb)
+        return FiniteVectorSpace.of(
+            scalars = ComplexAlgebras.ComplexField,
+            add = QuaternionDivisionRing.add,
+            dimension = 2,
+            leftAction = LeftAction { c, q ->
+                QuaternionDivisionRing.mul(embed(c), q)
+            }
+        )
+    }
+
+    // A map of all complex embeddings to the vector space that they generate.
+    val QuaternionLeftComplexVectorSpacesAll: Map<AxisSignEmbeddings.AxisSignEmbedding, FiniteVectorSpace<Complex, Quaternion>> =
+        AxisSignEmbeddings.AxisSignEmbedding.all.associateWith { quaternionLeftComplexVectorSpace(it) }
+
+    // Optional convenience vals, replacing the former AlongI/J/K trio.
+    val QuaternionLeftComplexVectorSpaceCanonical: FiniteVectorSpace<Complex, Quaternion> =
+        quaternionLeftComplexVectorSpace(canonicalEmbedding)
+
+    val QuaternionStarAlgebra: StarAlgebra<Real, Quaternion> = StarAlgebra.of(
+        scalars = RealAlgebras.RealField,
+        involutiveRing = QuaternionDivisionRing,
+        leftAction = QuaternionVectorSpace.leftAction
+    )
 
     /**
      * Return the ring monomorphism embedding ℂ into ℍ determined by [emb].
@@ -137,42 +166,11 @@ object QuaternionAlgebras {
     /**
      * Convenience extension.
      *
-     * Canonical default is `i_C ↦ i` (i.e. ComplexEmbedding.CANONICAL).
+     * Canonical default is `i_C ↦ i`.
      */
     fun Complex.asQuaternion(
         emb: AxisSignEmbeddings.AxisSignEmbedding = canonicalEmbedding
     ): Quaternion = complexEmbeddingToQuaternion(emb)(this)
-
-    /**
-     * ℍ as a left ℂ-vector space using the chosen embedding.
-     */
-    fun quaternionLeftComplexVectorSpace(
-        emb: AxisSignEmbeddings.AxisSignEmbedding = canonicalEmbedding
-    ): FiniteVectorSpace<Complex, Quaternion> {
-        val embed = complexEmbeddingToQuaternion(emb)
-        return FiniteVectorSpace.of(
-            scalars = ComplexAlgebras.ComplexField,
-            add = QuaternionDivisionRing.add,
-            dimension = 2,
-            leftAction = LeftAction { c, q ->
-                QuaternionDivisionRing.mul(embed(c), q)
-            }
-        )
-    }
-
-    // Optional convenience vals, replacing the former AlongI/J/K trio.
-    val QuaternionLeftComplexVectorSpaceCanonical: FiniteVectorSpace<Complex, Quaternion> =
-        quaternionLeftComplexVectorSpace(canonicalEmbedding)
-
-    // A map of all complex embeddings to the vector space that they generate.
-    val QuaternionLeftComplexVectorSpacesAll: Map<AxisSignEmbeddings.AxisSignEmbedding, FiniteVectorSpace<Complex, Quaternion>> =
-        AxisSignEmbeddings.AxisSignEmbedding.all.associateWith { quaternionLeftComplexVectorSpace(it) }
-
-    val QuaternionStarAlgebra: StarAlgebra<Real, Quaternion> = StarAlgebra.of(
-        scalars = RealAlgebras.RealField,
-        involutiveRing = QuaternionDivisionRing,
-        leftAction = QuaternionVectorSpace.leftAction
-    )
 
     val eqQuaternionStrict: Eq<Quaternion> = CD.eq(ComplexAlgebras.eqComplexStrict)
     val eqQuaternion: Eq<Quaternion> = CD.eq(ComplexAlgebras.eqComplex)
