+++ kosmos-core/src/main/kotlin/org/vorpal/kosmos/linear/ops/BLAS.kt
@@ -1,7 +1,7 @@
 package org.vorpal.kosmos.linear.ops
 
-import org.vorpal.kosmos.algebra.structures.instances.Complex
-import org.vorpal.kosmos.algebra.structures.instances.ComplexAlgebras
+import org.vorpal.kosmos.hypercomplex.complex.Complex
+import org.vorpal.kosmos.hypercomplex.complex.ComplexAlgebras
 import org.vorpal.kosmos.algebra.structures.instances.RealAlgebras
 import org.vorpal.kosmos.core.math.Real
 import org.vorpal.kosmos.linear.values.DenseMat
