+++ kosmos-testkit/src/main/kotlin/org/vorpal/kosmos/algebra/structures/instances/ArbComplex.kt
@@ -7,6 +7,7 @@ import io.kotest.property.arbitrary.map
 import io.kotest.property.arbitrary.pair
 import io.kotest.property.arbitrary.triple
 import org.vorpal.kosmos.algebra.structures.CD
+import org.vorpal.kosmos.hypercomplex.complex.Complex
 import kotlin.math.cos
 import kotlin.math.sin
